keyplayer: refactoring (step 1): use int/fixed point sample format
Reason for this change is that it will be more convenient to
represent phase using an fixed point type. This will facilitate
the implementation of phase modulation ("FM").
In particular, the phase wraps "naturally" (if using a 2PI range).

In this first step the present oscillator is kept + float-to-int
and int-to-float conversions ("E.Organ 1" still sounds right!)
The downside of this refactoring is that it will be impractical to
keep the Oscillator. It would have been nice to leverage on the
existing building blocks.