keyplayer: modulators had incorrect output level
This bug seems to explain why the patches tested so far sounded
disappointing. Whereas the carriers have to scale their output
levels to avoid clipping, the modulators shouldn't (but they did!).
The net effect was that the programs sounded mellow and dull, since
the modulation index was too low.
This fix makes a huge improvement of the sound quality!