keyplayer: SysEx transfer of Packed 32 Voice format (DX7 .syx)
Due to a limitation in MidiHandler, the .syx files need to be
chopped up in fragments of max 128 bytes. The Packed 32 Voice
format is 4104 bytes in size (including first 0xf0 and last 0xf7),
which means that the message needs to be chopped up in at least
33 parts.

Removed the preloaded program bank again. The (rather dull) init
voice all you got out-of-the-box, before loading any .syx file.