op6daisy: use look-up table for sine wave (step 1)
The loop in FmOperator::fillbuffer() is the hottest code in the
whole program. The loop body is executed 288k times per second
and voice. To get much further than four voices of polyphony,
it has to be optimized.

To get this right, I make this change in a stepwise fashion.

Step 1 changes the data type of the instantaneous phase modulation
(m) from float to 32-bit int. [-2^31, +2^31) maps onto the real
range [-PI,+PI].