op6daisy: use look-up table for sine wave (step 2)
The loop in FmOperator::fillbuffer() is the hottest code in the
whole program. The loop body is executed 288k times per second
and voice. To get much further than four voices of polyphony,
it has to be optimized.

To get this right, I make this change in a stepwise fashion.

Step 2 changes the data type of the delayed samples from float to
int. The samples are stored in 24-bit fixed-point format (Q24),
which turns out to fit the use (in the following loop iteration)
well. All that is needed to compute the modulator is an addition
and a shift.