op6daisy: use look-up table for sine wave (step 3)
The loop in FmOperator::fillbuffer() is the hottest code in the
whole program. The loop body is executed 288k times per second
and voice. To get much further than four voices of polyphony,
it has to be optimized.

To get this right, I make this change in a stepwise fashion.

Step 3 introduces a dummy function for the look-up table -it still
uses the sinf() libcall.