daisychain: sync to external clock
The original code (from Daisy Patch) is intended for an external
codec only, in which case the Daisy always is the clock master.
By this adaptation it is possible to daisy chain two Daisies such
that one of them syncs to the clock generated by the other device.

The state of pin D15 at startup controls the role of the Daisy.
Unconnected (pulled up) or HIGH = clock master. If grounded/LOW,
the Daisy syncs to external clock.