op6daisy: increased NUM_VOICES to 16 (was 4)
After optimizing the hot loop it seems safe to increase polyphony
to 16. This means it is now possible to play 16 notes the same
time: all on a single MIDI channel or distributed over several
channels.

This hasn't been tested rigorously. The LED, which flashes when
MIDI note-on events are received, is lit up permanently if there
is a *buffer underrun*. An underrun means that the code fails to
deliver the audio samples in a timely fashion. The higher the
polyphony, the tougher it is to meet the deadlines. As long as
that LED doesn't signal underrun, op6daisy is managing well.
So, you might want to keep an eye on that LED.