op6daisy: don't retrigger envelope when recycling a voice
Fast retriggering of notes (ie same note/key) sometimes caused
clicking sounds. Particularly so for sustained sounds.

The voice-stealing algorithms favours stealing a possible active
voice for the channel/key over stealing the voice of another
channel or key (for most instruments it doesn't make sense to
have several active instances of the same note). This means that
the recycled voice might not have decayed, which makes an abrupt
retriggering cause a clicking artifact.

It seems the issue has been solved (or at least mitigated) by not
retriggering the envelopes when recycling an active voice for a
repeated note on the same channel.