op6daisy: Manage Temporary Programs in Voice Edit Mode
When auditing changes in Voice Edit Mode (playing notes), we can't
use the Program that is being edited. Instead a collection of
temporary programs are maintained. In the worst case, we need as
many temporary programs as there are voices. Reference counting is
employed, such that we know when a temporary program can be reused.