op6-ui: the voice-editor controller stores the whole program bank
The voice editor used to rely on the loaded .syx file for the
voice parameters, they were converted from .syx when opening the
voice editor. Now, the .syx file is instead converted when loading
the program bank and the voice-editor controller manages these
parameters. This means there is a place to store voices that are
updated by the editor (beyond the single voice that is currently
being edited).