op6daisy: monophonic operation mode
MIDI CC #126 (mono) and #127 (poly) control operation mode.
Separate modes are maintained per MIDI channel, such that there
may be 16 parts in individual operation modes.
A more common approach is that the instrument is in one of four
MIDI modes, which is controlled by mono/poly and omni on/off
messages.