op6daisy: keyboard priority in monophonic mode
In monophonic mode only one key produces sound. If several keys
are pressed and held, the one triggered last is the one sounding.
When it is released, one of the other keys take precedence
according to keyboard priority. The keyboard priority depends on
the two most recent keys played:
a) if a first note is followed by a second, higher note, *high
   note priority* is used. When the second note is released,
   the highest note still held then takes precedence.
b) if a first note is followed by a second, lower note, *low
   note priority* is used. When the second note is released,
   the lowest note still held then takes precedence.
The new note is played legato, with no re-triggering of envelopes.